[Engine][ResourceManager] move owned slot profile from scheduler to job master.
