+++ seatunnel-engine/seatunnel-engine-server/src/main/java/org/apache/seatunnel/engine/server/master/JobMaster.java
@@ -29,11 +29,13 @@ import org.apache.seatunnel.engine.server.checkpoint.CheckpointManager;
 import org.apache.seatunnel.engine.server.checkpoint.CheckpointPlan;
 import org.apache.seatunnel.engine.server.checkpoint.CheckpointStorageConfiguration;
 import org.apache.seatunnel.engine.server.dag.physical.PhysicalPlan;
+import org.apache.seatunnel.engine.server.dag.physical.PhysicalVertex;
 import org.apache.seatunnel.engine.server.dag.physical.PlanUtils;
 import org.apache.seatunnel.engine.server.resourcemanager.ResourceManager;
-import org.apache.seatunnel.engine.server.scheduler.JobScheduler;
+import org.apache.seatunnel.engine.server.resourcemanager.resource.SlotProfile;
 import org.apache.seatunnel.engine.server.scheduler.PipelineBaseScheduler;
 
+import com.hazelcast.cluster.Address;
 import com.hazelcast.flakeidgen.FlakeIdGenerator;
 import com.hazelcast.internal.serialization.Data;
 import com.hazelcast.jet.datamodel.Tuple2;
@@ -45,7 +47,9 @@ import lombok.NonNull;
 import org.apache.commons.collections4.CollectionUtils;
 
 import java.util.Map;
+import java.util.Optional;
 import java.util.concurrent.CompletableFuture;
+import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.ExecutorService;
 
 public class JobMaster implements Runnable {
@@ -69,7 +73,7 @@ public class JobMaster implements Runnable {
 
     private JobImmutableInformation jobImmutableInformation;
 
-    private JobScheduler jobScheduler;
+    private final Map<Integer, Map<PhysicalVertex, SlotProfile>> ownedSlotProfiles;
 
     public JobMaster(@NonNull Data jobImmutableInformationData,
                      @NonNull NodeEngine nodeEngine,
@@ -79,7 +83,7 @@ public class JobMaster implements Runnable {
         this.executorService = executorService;
         flakeIdGenerator =
             this.nodeEngine.getHazelcastInstance().getFlakeIdGenerator(Constant.SEATUNNEL_ID_GENERATOR_NAME);
-
+        this.ownedSlotProfiles = new ConcurrentHashMap<>();
         this.resourceManager = resourceManager;
     }
 
@@ -129,9 +133,7 @@ public class JobMaster implements Runnable {
                 }
                 jobMasterCompleteFuture.complete(physicalPlan.getJobStatus());
             });
-
-            jobScheduler = new PipelineBaseScheduler(physicalPlan, this);
-            jobScheduler.startScheduling();
+            ownedSlotProfiles.putAll(new PipelineBaseScheduler(physicalPlan, this).startScheduling());
         } catch (Throwable e) {
             LOGGER.severe(String.format("Job %s (%s) run error with: %s",
                 physicalPlan.getJobImmutableInformation().getJobConfig().getName(),
@@ -148,8 +150,16 @@ public class JobMaster implements Runnable {
         // TODO Add some job clean operation
     }
 
-    public JobScheduler getJobScheduler() {
-        return jobScheduler;
+    public Address queryTaskGroupAddress(long taskGroupId) {
+        for (Integer pipelineIndex : ownedSlotProfiles.keySet()) {
+            Optional<PhysicalVertex> currentVertex = ownedSlotProfiles.get(pipelineIndex).keySet().stream()
+                .filter(physicalVertex -> physicalVertex.getTaskGroup().getTaskGroupLocation().getTaskGroupId() == taskGroupId)
+                .findFirst();
+            if (currentVertex.isPresent()) {
+                return ownedSlotProfiles.get(pipelineIndex).get(currentVertex.get()).getWorker();
+            }
+        }
+        throw new IllegalArgumentException("can't find task group address from task group id: " + taskGroupId);
     }
 
     public void cancelJob() {
